WIP Still Making GameList action when Clicked game item.
